# Improve performance of UriHelper.GetDisplayUrl