@amcasey 2024-05-15
I'm going to be sad if this is the most efficient way to concatenate six strings.