@paulomorgado 2024-05-16
> I'm going to be sad if this is the most efficient way to concatenate six strings.

Up to 9.0.0-preview.3.24175.3, it is.