@paulomorgado 2024-05-21
Upated to use the new in **.NET 9.0** `String.Concat(ReadOnlySpan<string?> values)` method.

Always slower than `String.Create` and slower than string interpolation in most of the cases.