@stephentoub 2024-06-02
> Always slower than String.Create and slower than string interpolation in most of the cases.

I don't see a difference anywhere close to what you're seeing. For me, the string.Create case is just a few ns faster than the string.Concat (which makes sense, as it's effectively manually unrolling a loop the latter has), but it's also ~50 lines of code compared to the ~1 line of code for string.Concat. And it's going to contribute to a larger NativeAOT footprint, due to the use of the ValueTuple`5, which may not be otherwise used in the app and which brings with it a non-trivial footprint.  This path is already allocating, so it can't be an ultra hot path. (I also wouldn't be surprised if the string.Create benchmark were being skewed a bit by dynamic PGO, which can devirtualize this delegate call but would be much less likely in a real app where different delegates were being used with it.)

I'd just do the super simple and basically as efficient thing; just use Concat.