@paulomorgado 2024-06-02
Hi @stephentoub,

> > Always slower than String.Create and slower than string interpolation in most of the cases.
> 
> I don't see a difference anywhere close to what you're seeing. For me, the string.Create case is just a few ns faster than the string.Concat (which makes sense, as it's effectively manually unrolling a loop the latter has), but it's also ~50 lines of code compared to the ~1 line of code for string.Concat. And it's going to contribute to a larger NativeAOT footprint, due to the use of the ValueTuple`5, which may not be otherwise used in the app and which brings with it a non-trivial footprint. (I also wouldn't be surprised if the string.Create benchmark were being skewed a bit by dynamic PGO, which can devirtualize this delegate call but would be much less likely in a real app where different delegates were being used with it.)
> 
> This path is already allocating, so it can't be an ultra hot path. I'd just do the super simple and basically as efficient thing; just use Concat.

I was comparing the current implementation using `StringBuilder`.
This PR is based on #28906 from Dec 29, 2020 and I was only made aware of the existence of `String.Concat(ReadOnlySpan<string?> values)` in **.NET 9.0** by @MihaZupan's [comment](https://github.com/dotnet/aspnetcore/pull/55611#discussion_r1595520503).

In the meantime the original idea of using `string.Create` was picked [opentelemetry-dotnet](https://github.com/open-telemetry/opentelemetry-dotnet) picked up the original idea on https://github.com/open-telemetry/opentelemetry-dotnet/pull/2947 and [ImageSharp.Web](https://github.com/SixLabors/ImageSharp.Web) on https://github.com/SixLabors/ImageSharp.Web/pull/265.

In the interest of measuring and comparing, I've added:

```csharp
[Benchmark]
[ArgumentsSource(nameof(Data))]
public string String_Concat2(string scheme, HostString host, PathString basePath, PathString path, QueryString query)
{
    if (!query.HasValue)
    {
        return string.Concat(scheme, "://", host.Value, basePath.Value, path.Value);
    }
    else if (!basePath.HasValue)
    {
        return string.Concat(scheme, "://", host.Value, path.Value, query.Value);
    }
    else if (!path.HasValue)
    {
        return string.Concat(scheme, "://", host.Value, basePath.Value, query.Value);
    }
    else
    {
        return string.Concat((ReadOnlySpan<string>)[scheme, "://", host.Value, basePath.Value, path, query.Value]);
    }
}
```

But it doesn't perform better:

| Method               | scheme | host             | basePath   | path                | query                | Mean      | Gen0   | Allocated |
|--------------------- |------- |----------------- |----------- |-------------------- |--------------------- |----------:|-------:|----------:|
| String_Concat        | http   | cname.domain.tld |            | /                   |                      | 22.490 ns | 0.0038 |      72 B |
| String_Concat2       | http   | cname.domain.tld |            | /                   |                      | 19.873 ns | 0.0072 |     136 B |
|                      |        |                  |            |                     |                      |           |        |           |
| String_Concat        | http   | cname.domain.tld |            | /                   | ?para(...)alue3 [42] | 24.928 ns | 0.0085 |     160 B |
| String_Concat2       | http   | cname.domain.tld |            | /                   | ?para(...)alue3 [42] | 22.459 ns | 0.0119 |     224 B |
|                      |        |                  |            |                     |                      |           |        |           |
| String_Concat        | http   | cname.domain.tld |            | /path/one/two/three |                      | 34.783 ns | 0.0059 |     112 B |
| String_Concat2       | http   | cname.domain.tld |            | /path/one/two/three |                      | 21.974 ns | 0.0093 |     176 B |
|                      |        |                  |            |                     |                      |           |        |           |
| String_Concat        | http   | cname.domain.tld |            | /path/one/two/three | ?para(...)alue3 [42] | 44.334 ns | 0.0102 |     192 B |
| String_Concat2       | http   | cname.domain.tld |            | /path/one/two/three | ?para(...)alue3 [42] | 24.625 ns | 0.0136 |     256 B |
|                      |        |                  |            |                     |                      |           |        |           |
| String_Concat        | http   | cname.domain.tld | /base-path | /                   |                      | 20.708 ns | 0.0051 |      96 B |
| String_Concat2       | http   | cname.domain.tld | /base-path | /                   |                      | 24.585 ns | 0.0085 |     160 B |
|                      |        |                  |            |                     |                      |           |        |           |
| String_Concat        | http   | cname.domain.tld | /base-path | /                   | ?para(...)alue3 [42] | 23.504 ns | 0.0093 |     176 B |
| String_Concat2       | http   | cname.domain.tld | /base-path | /                   | ?para(...)alue3 [42] | 28.154 ns | 0.0093 |     176 B |
|                      |        |                  |            |                     |                      |           |        |           |
| String_Concat        | http   | cname.domain.tld | /base-path | /path/one/two/three |                      | 36.179 ns | 0.0068 |     128 B |
| String_Concat2       | http   | cname.domain.tld | /base-path | /path/one/two/three |                      | 23.420 ns | 0.0102 |     192 B |
|                      |        |                  |            |                     |                      |           |        |           |
| String_Concat        | http   | cname.domain.tld | /base-path | /path/one/two/three | ?para(...)alue3 [42] | 41.275 ns | 0.0114 |     216 B |
| String_Concat2       | http   | cname.domain.tld | /base-path | /path/one/two/three | ?para(...)alue3 [42] | 44.670 ns | 0.0114 |     216 B |

I haven't doen't any work with AOT, so that completely overlooked that.

But, I tried creating a struct for the arguments:

```csharp
[Benchmark]
[ArgumentsSource(nameof(Data))]
public string String_Create2(string scheme, HostString host, PathString basePath, PathString path, QueryString query)
{
    var schemeValue = scheme ?? string.Empty;
    var hostValue = host.Value ?? string.Empty;
    var basePathValue = basePath.Value ?? string.Empty;
    var pathValue = path.Value ?? string.Empty;
    var queryValue = query.Value ?? string.Empty;

    var length =
        +schemeValue.Length
        + SchemeDelimiter.Length
        + hostValue.Length
        + basePathValue.Length
        + pathValue.Length
        + queryValue.Length;

    return string.Create(
        length,
        new UriParts( scheme, hostValue, basePathValue, pathValue, queryValue),
        static (buffer, uriParts) =>
        {
            if (uriParts.Scheme.Length > 0)
            {
                uriParts.Scheme.CopyTo(buffer);
                buffer = buffer.Slice(uriParts.Scheme.Length);
            }

            SchemeDelimiter.CopyTo(buffer);
            buffer = buffer.Slice(SchemeDelimiter.Length);

            if (uriParts.Host.Length > 0)
            {
                uriParts.Host.CopyTo(buffer);
                buffer = buffer.Slice(uriParts.Host.Length);
            }

            if (uriParts.BasePath.Length > 0)
            {
                uriParts.BasePath.CopyTo(buffer);
                buffer = buffer.Slice(uriParts.BasePath.Length);
            }

            if (uriParts.Path.Length > 0)
            {
                uriParts.Path.CopyTo(buffer);
                buffer = buffer.Slice(uriParts.Path.Length);
            }

            if (uriParts.Query.Length > 0)
            {
                uriParts.Query.CopyTo(buffer);
            }
        });
}

private readonly struct UriParts
{
    public readonly string Scheme;
    public readonly string Host;
    public readonly string BasePath;
    public readonly string Path;
    public readonly string Query;

    public UriParts(string scheme, string host, string basePath, string path, string query)
    {
        this.Scheme = scheme;
        this.Host = host;
        this.BasePath = basePath;
        this.Path = path;
        this.Query = query;
    }
}
```

And it's very close that using a tuple (`DOTNET_TieredPGO` set to `0`):

| Method               | scheme | host             | basePath   | path                | query                | Mean     | Gen0   | Allocated |
|--------------------- |------- |----------------- |----------- |-------------------- |--------------------- |---------:|-------:|----------:|
| String_Concat        | http   | cname.domain.tld |            | /                   |                      | 22.50 ns | 0.0038 |      72 B |
| String_Create        | http   | cname.domain.tld |            | /                   |                      | 10.13 ns | 0.0038 |      72 B |
| String_Create2       | http   | cname.domain.tld |            | /                   |                      | 12.72 ns | 0.0038 |      72 B |
|                      |        |                  |            |                     |                      |          |        |           |
| String_Concat        | http   | cname.domain.tld |            | /                   | ?para(...)alue3 [42] | 27.57 ns | 0.0085 |     160 B |
| String_Create        | http   | cname.domain.tld |            | /                   | ?para(...)alue3 [42] | 16.74 ns | 0.0085 |     160 B |
| String_Create2       | http   | cname.domain.tld |            | /                   | ?para(...)alue3 [42] | 14.38 ns | 0.0085 |     160 B |
|                      |        |                  |            |                     |                      |          |        |           |
| String_Concat        | http   | cname.domain.tld |            | /path/one/two/three |                      | 45.80 ns | 0.0059 |     112 B |
| String_Create        | http   | cname.domain.tld |            | /path/one/two/three |                      | 12.34 ns | 0.0059 |     112 B |
| String_Create2       | http   | cname.domain.tld |            | /path/one/two/three |                      | 11.40 ns | 0.0059 |     112 B |
|                      |        |                  |            |                     |                      |          |        |           |
| String_Concat        | http   | cname.domain.tld |            | /path/one/two/three | ?para(...)alue3 [42] | 49.75 ns | 0.0102 |     192 B |
| String_Create        | http   | cname.domain.tld |            | /path/one/two/three | ?para(...)alue3 [42] | 16.61 ns | 0.0102 |     192 B |
| String_Create2       | http   | cname.domain.tld |            | /path/one/two/three | ?para(...)alue3 [42] | 15.46 ns | 0.0102 |     192 B |
|                      |        |                  |            |                     |                      |          |        |           |
| String_Concat        | http   | cname.domain.tld | /base-path | /                   |                      | 23.96 ns | 0.0051 |      96 B |
| String_Create        | http   | cname.domain.tld | /base-path | /                   |                      | 13.31 ns | 0.0051 |      96 B |
| String_Create2       | http   | cname.domain.tld | /base-path | /                   |                      | 11.28 ns | 0.0051 |      96 B |
|                      |        |                  |            |                     |                      |          |        |           |
| String_Concat        | http   | cname.domain.tld | /base-path | /                   | ?para(...)alue3 [42] | 27.16 ns | 0.0093 |     176 B |
| String_Create        | http   | cname.domain.tld | /base-path | /                   | ?para(...)alue3 [42] | 17.88 ns | 0.0093 |     176 B |
| String_Create2       | http   | cname.domain.tld | /base-path | /                   | ?para(...)alue3 [42] | 14.86 ns | 0.0093 |     176 B |
|                      |        |                  |            |                     |                      |          |        |           |
| String_Concat        | http   | cname.domain.tld | /base-path | /path/one/two/three |                      | 45.87 ns | 0.0068 |     128 B |
| String_Create        | http   | cname.domain.tld | /base-path | /path/one/two/three |                      | 14.12 ns | 0.0068 |     128 B |
| String_Create2       | http   | cname.domain.tld | /base-path | /path/one/two/three |                      | 12.51 ns | 0.0068 |     128 B |
|                      |        |                  |            |                     |                      |          |        |           |
| String_Concat        | http   | cname.domain.tld | /base-path | /path/one/two/three | ?para(...)alue3 [42] | 50.87 ns | 0.0114 |     216 B |
| String_Create        | http   | cname.domain.tld | /base-path | /path/one/two/three | ?para(...)alue3 [42] | 19.09 ns | 0.0115 |     216 B |
| String_Create2       | http   | cname.domain.tld | /base-path | /path/one/two/three | ?para(...)alue3 [42] | 16.46 ns | 0.0115 |     216 B |

Whether this marginally improvement over `string.Concat` is worst all that code, would require data I don't have on the usage of this API.

The best option would be to add an overload to `string.Concat` with 5 `string` parameters, but that would be a breaking change. One that might not hurt anyone, but still a breaking change.